requirements.txt
streamlit
python-dotenv
azure-identity
azure-ai-projects
PyPDF2
agent-framework
PyMuPDF